:recycle: Refactor polymorphic association for sticky
According to the rule of RoR, describe polymorphic association for sticky
model more accurately. Sticky.target_type must be model name in this rule,
so target_type values are also changed with 'PageFile' and ' Note'.
If some records already exist in the sticky table, it is necessary to modify
the target_type values of the records.